TrafficOptima/Monitoring_System/Server_Application/Flask_Server/requirements.txt
absl-py==1.4.0
astroid==2.15.6
astunparse==1.6.3
audioread==3.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
dill==0.3.7
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.42.1
Flask-Cors==3.0.9
Flask-Bcrypt==0.7.1
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
idna==3.4
imageio==2.31.2
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.40.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
numba==0.57.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
platformdirs==3.10.0
pooch==1.7.0
protobuf==4.24.2
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pylint==2.17.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
PyJWT==1.5.3
pymongo==3.10.1
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
soundfile==0.12.1
soxr==0.3.6
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.8.25
tomlkit==0.12.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.163
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
